Colt suggestion/multi tenancy acls (#145)
* refactor the proto to allow for:
- one Principal has different ACL's in different Tenant's
- allow for setting 'cluster-wide' permissions for a Principal

* adds C# option for acl proto

* Updates:

- updates the PrincipalModel to follow the correct proto
- Makes PutTenantRequest and PutPrincipalRequest return Tenant and Principal, which follows pattern of PutTaskDef/PutWfSpec etc
- Improves validation of PutPrincipalRequest
- Fixes createdAt on Tenant and Principal

* - Cleans up the PutPrincipalRequest and PutTenantRequest.
- Remove old commented code in CoreProcessorDAO.

Next up:
- Re-enable onPartitionClaimed() on the core processor DAO (which means: send a message to yourself). Will be used to safeguard RYOW during a rebalance.
- Make the CoreProcessorDAO and the MetadataProcessorDAO share a common abstract class so that we don't have to duplicate code around managing Tags.

* - Now deletes tags of old getable's in the Metadata Processor
- Fixes many warnings about missing cases in switch, unused private variables, unchecked casts, etc.

* fix more missing case statements and unused variables

* Renames ReadOnlyMetadataProcessorDAO to ReadOnlyMetadataDAO since it is used in the GRPC Query Server (interactive queries) and not only in the processor

* adds getPrincipal method to the dao

* un-does the renaming of ReadOnlyMetadataProcessorDAO -> ReadOnlyMetadataDAO in order to reduce size of PR

+++ server/src/main/java/io/littlehorse/common/model/metadatacommand/subcommand/DeletePrincipalRequestModel.java
@@ -1,16 +1,20 @@
 package io.littlehorse.common.model.metadatacommand.subcommand;
 
+import com.google.protobuf.Empty;
 import com.google.protobuf.Message;
+import io.grpc.Status;
+import io.littlehorse.common.LHConstants;
 import io.littlehorse.common.LHServerConfig;
 import io.littlehorse.common.dao.MetadataProcessorDAO;
-import io.littlehorse.common.model.ServerSubCommand;
+import io.littlehorse.common.exceptions.LHApiException;
+import io.littlehorse.common.model.ClusterLevelCommand;
 import io.littlehorse.common.model.getable.objectId.PrincipalIdModel;
 import io.littlehorse.common.model.metadatacommand.MetadataSubCommand;
 import io.littlehorse.common.proto.DeletePrincipalRequest;
 import io.littlehorse.sdk.common.exception.LHSerdeError;
 
 public class DeletePrincipalRequestModel extends MetadataSubCommand<DeletePrincipalRequest>
-        implements ServerSubCommand {
+        implements ClusterLevelCommand {
 
     private String id;
 
@@ -36,8 +40,15 @@ public class DeletePrincipalRequestModel extends MetadataSubCommand<DeletePrinci
     }
 
     @Override
-    public Message process(MetadataProcessorDAO dao, LHServerConfig config) {
+    public Empty process(MetadataProcessorDAO dao, LHServerConfig config) {
+        if (id.equals(LHConstants.ANONYMOUS_PRINCIPAL)) {
+            throw new LHApiException(
+                    Status.INVALID_ARGUMENT,
+                    "Deleting the anonymous principal is not allowed. However, "
+                            + "you can remove its permissions via the PutPrincipal request.");
+        }
+
         dao.delete(new PrincipalIdModel(id));
-        return null;
+        return Empty.getDefaultInstance();
     }
 }
